requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
Django==1.11.17
django-rq==2.1.0
djangorestframework==3.11.0
idna==2.9
psycopg2==2.8.5
pytz==2020.1
redis==3.4.1
requests==2.23.0
rq==1.3.0
sqlparse==0.3.1
urllib3==1.25.9
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.17
django-rq==2.1.0
djangorestframework==3.11.0
drf-yasg==1.17.1
idna==2.9
inflection==0.4.0
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
packaging==20.3
psycopg2==2.8.5
pyparsing==2.4.7
pytz==2020.1
redis==3.4.1
requests==2.23.0
rq==1.3.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.14.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
